Refactor Index page components
Refactored the `src/pages/Index.tsx` component into smaller, more maintainable components for the Hero, Features, and How It Works sections.
